refactor(core): promote pkg/updater to internal/core/updater (no behavior change)
