+++ internal/core/updater/notice_test.go
